chore: move accordion block to the correct folder
