+++ libs/blocks/src/index.ts
@@ -1,4 +1,4 @@
-export * from './accordion';
+export * from './lib/accordion';
 export * from './lib/content-text';
 export * from './lib/account-comparison';
 export * from './lib/stat';
